start of breaking precise into sp3 and clk
