+++ .gitignore
@@ -42,10 +42,6 @@ results/*
 *.*g
 *.rnx
 # Include unit test files for broadcast ephemerides
-!data/unit_test/*/*.*n
-!data/unit_test/*/*.*o
-!data/unit_test/*/*.*g
-!data/unit_test/*/*.rnx
 !data/unit_test/*/*/*.*n
 !data/unit_test/*/*/*.*o
 !data/unit_test/*/*/*.*g
