+++ gnss_lib_py/utils/sv_models.py
@@ -795,3 +795,275 @@ def _compute_eccentric_anomaly(gps_week, gps_tow, ephem, tol=1e-5, max_iter=10):
                             + f"after {max_iter} steps. : dE = {delta_ecc_anom}")
 
     return ecc_anom
+
+
+def single_gnss_from_precise_eph(navdata, sp3_parsed_file,
+                                 clk_parsed_file, inplace=False,
+                                 verbose = False):
+    """Compute satellite information using .sp3 and .clk for any GNSS constellation
+
+    Either adds or replaces satellite ECEF position data and clock bias
+    for any satellite that exists in the provided sp3 file
+
+    Parameters
+    ----------
+    navdata : gnss_lib_py.parsers.navdata.NavData
+        Instance of the NavData class that depicts android derived dataset
+    sp3_parsed_file : list
+        Instance with list of gnss_lib_py.parsers.precise_ephemerides.Sp3
+    clk_parsed_file : list
+        Instance with list of gnss_lib_py.parsers.precise_ephemerides.Clk
+    inplace : bool
+        If true, adds satellite positions and clock bias to the input
+        navdata object, otherwise returns a new NavData object with the
+        satellite rows added.
+    verbose : bool
+        Flag (True/False) for whether to print intermediate steps useful
+        for debugging/reviewing (the default is False)
+
+    Returns
+    -------
+    navdata : gnss_lib_py.parsers.navdata.NavData
+        Updated NavData class with satellite information computed using
+        precise ephemerides from .sp3 and .clk files
+    """
+
+    # Initialize the sp3 and clk iref arrays
+    sp3_iref_old = {}
+    satfunc_xyz_old = {}
+    clk_iref_old = {}
+    satfunc_t_old = {}
+
+
+    # combine gnss_id and sv_id into gnss_sv_ids
+    if inplace:
+        navdata["gnss_sv_id"] = _combine_gnss_sv_ids(navdata)
+    else:
+        new_navdata = navdata.copy()
+        new_navdata["gnss_sv_id"] = _combine_gnss_sv_ids(new_navdata)
+
+    # add satellite indexes if not present already.
+    sv_idx_keys = ['x_sv_m', 'y_sv_m', 'z_sv_m', \
+                'vx_sv_mps','vy_sv_mps','vz_sv_mps', \
+                'b_sv_m', 'b_dot_sv_mps']
+    for sv_idx_key in sv_idx_keys:
+        if sv_idx_key not in navdata.rows:
+            if inplace:
+                navdata[sv_idx_key] = np.nan
+            else:
+                new_navdata[sv_idx_key] = np.nan
+
+
+    if inplace:
+        iterate_navdata = navdata
+    else:
+        iterate_navdata = new_navdata
+
+    for row_idx, row in enumerate(iterate_navdata):
+        gnss_sv_id = str(row["gnss_sv_id"])
+        # continue if no sp3 or clk data availble
+        if gnss_sv_id not in sp3_parsed_file \
+          or gnss_sv_id not in clk_parsed_file \
+          or len(sp3_parsed_file[gnss_sv_id].tym) == 0 \
+          or len(clk_parsed_file[gnss_sv_id].tym) == 0: continue
+
+        timestep = row["gps_millis"]
+
+        # Perform nearest time step search to compute iref values for sp3 and clk
+        sp3_iref = np.argmin(abs(np.array(sp3_parsed_file[gnss_sv_id].tym) - \
+                                 timestep ))
+        clk_iref = np.argmin(abs(np.array(clk_parsed_file[gnss_sv_id].tym) - \
+                                 timestep ))
+
+        # Carry out .sp3 processing by first checking if
+        # previous interpolated function holds
+        if gnss_sv_id in sp3_iref_old and sp3_iref == sp3_iref_old[gnss_sv_id]:
+            func_satpos = satfunc_xyz_old[gnss_sv_id]
+        else:
+            # if does not hold, recompute the interpolation function based on current iref
+            if verbose:
+                print('SP3: Computing new interpolation for',gnss_sv_id)
+            func_satpos = extract_sp3(sp3_parsed_file[gnss_sv_id], \
+                                           sp3_iref)
+            # Update the relevant interp function and iref values
+            satfunc_xyz_old[gnss_sv_id] = func_satpos
+            sp3_iref_old[gnss_sv_id] = sp3_iref
+
+        # Compute satellite position and velocity using interpolated function
+        satpos_sp3, satvel_sp3 = sp3_snapshot(func_satpos, timestep)
+
+        # Adjust the satellite position based on Earth's rotation
+        trans_time = row["raw_pr_m"] / consts.C
+        del_x = (consts.OMEGA_E_DOT * satpos_sp3[1] * trans_time)
+        del_y = (-consts.OMEGA_E_DOT * satpos_sp3[0] * trans_time)
+        satpos_sp3[0] = satpos_sp3[0] + del_x
+        satpos_sp3[1] = satpos_sp3[1] + del_y
+
+        # Carry out .clk processing by first checking if previous interpolated
+        # function holds
+        if gnss_sv_id in clk_iref_old and clk_iref == clk_iref_old[gnss_sv_id]:
+            func_satbias = satfunc_t_old[gnss_sv_id]
+        else:
+            # if does not hold, recompute the interpolation function based on current iref
+            if verbose:
+                print('CLK: Computing new interpolation for',gnss_sv_id)
+            func_satbias = extract_clk(clk_parsed_file[gnss_sv_id],
+                                            clk_iref)
+            # Update the relevant interp function and iref values
+            satfunc_t_old[gnss_sv_id] = func_satbias
+            clk_iref_old[gnss_sv_id] = clk_iref
+
+        # Compute satellite clock bias and drift using interpolated function
+        satbias_clk, satdrift_clk = clk_snapshot(func_satbias, timestep)
+
+        if inplace:
+            # update *_sv_m of navdata with the estimated values from .sp3 files
+            navdata['x_sv_m', row_idx] = np.array([satpos_sp3[0]])
+            navdata['y_sv_m', row_idx] = np.array([satpos_sp3[1]])
+            navdata['z_sv_m', row_idx] = np.array([satpos_sp3[2]])
+
+            # update v*_sv_mps of navdata with the estimated values from .sp3 files
+            navdata["vx_sv_mps", row_idx] = np.array([satvel_sp3[0]])
+            navdata["vy_sv_mps", row_idx] = np.array([satvel_sp3[1]])
+            navdata["vz_sv_mps", row_idx] = np.array([satvel_sp3[2]])
+
+            # update clock data of navdata with the estimated values from .clk files
+            navdata["b_sv_m", row_idx] = np.array([consts.C * satbias_clk])
+            navdata["b_dot_sv_mps", row_idx] = np.array([consts.C * satdrift_clk])
+        else:
+            # update *_sv_m of navdata with the estimated values from .sp3 files
+            new_navdata['x_sv_m', row_idx] = np.array([satpos_sp3[0]])
+            new_navdata['y_sv_m', row_idx] = np.array([satpos_sp3[1]])
+            new_navdata['z_sv_m', row_idx] = np.array([satpos_sp3[2]])
+
+            # update v*_sv_mps of navdata with the estimated values from .sp3 files
+            new_navdata["vx_sv_mps", row_idx] = np.array([satvel_sp3[0]])
+            new_navdata["vy_sv_mps", row_idx] = np.array([satvel_sp3[1]])
+            new_navdata["vz_sv_mps", row_idx] = np.array([satvel_sp3[2]])
+
+            # update clock data of navdata with the estimated values from .clk files
+            new_navdata["b_sv_m", row_idx] = np.array([consts.C * satbias_clk])
+            new_navdata["b_dot_sv_mps", row_idx] = np.array([consts.C * satdrift_clk])
+
+    if inplace:
+        return None
+    return new_navdata
+
+def add_sv_states_sp3_and_clk(navdata, sp3_path, clk_path,
+                                inplace=False, verbose = False):
+    """Compute satellite information using .sp3 and .clk for multiple GNSS
+
+    Parameters
+    ----------
+    navdata : gnss_lib_py.parsers.navdata.NavData
+        Instance of the NavData class that depicts android derived dataset
+    sp3_path : path
+        File path for .sp3 file to extract precise ephemerides
+    clk_path : path
+        File path for .clk file to extract precise ephemerides
+    inplace : bool
+        If true, adds satellite positions and clock bias to the input
+        navdata object, otherwise returns a new NavData object with the
+        satellite rows added.
+    verbose : bool
+        Flag for whether to print intermediate steps useful
+        for debugging/reviewing (the default is False)
+
+    Returns
+    -------
+    navdata : gnss_lib_py.parsers.navdata.NavData
+        Updated NavData class with satellite information computed using
+        precise ephemerides from .sp3 and .clk files
+    """
+    sp3_parsed_gnss = parse_sp3(sp3_path)
+    clk_parsed_gnss = parse_clockfile(clk_path)
+    precise_navdata = single_gnss_from_precise_eph(navdata,
+                                                   sp3_parsed_gnss,
+                                                   clk_parsed_gnss,
+                                                   inplace = inplace,
+                                                   verbose = verbose)
+
+    return precise_navdata
+
+def sv_gps_from_brdcst_eph_duplicate(navdata, verbose = False):
+    """Compute satellite information using .n for any GNSS constellation
+
+    Parameters
+    ----------
+    navdata : gnss_lib_py.parsers.navdata.NavData
+        Instance of the NavData class that depicts android derived dataset
+    verbose : bool
+        Flag (True/False) for whether to print intermediate steps useful
+        for debugging/reviewing (the default is False)
+
+    Returns
+    -------
+    navdata : gnss_lib_py.parsers.navdata.NavData
+        Updated NavData class with satellite information computed using
+        broadcast ephemerides from .n files
+    """
+    unique_gnss_id = np.unique(navdata['gnss_id'])
+    if len(unique_gnss_id)==1:
+        if unique_gnss_id == 'gps':
+            # Need this string to create sv_id strings for ephemeris manager
+            unique_gnss_id_str = 'G'
+        else:
+            raise RuntimeError("No non-GPS capability yet")
+    else:
+        raise RuntimeError("Multi-GNSS constellations cannot be updated simultaneously")
+
+    unique_timesteps = np.unique(navdata["gps_millis"])
+
+    for _, timestep in enumerate(unique_timesteps):
+        # Compute indices where gps_millis match, sort them
+        # sorting is done for consistency across all satellite pos. estimation
+        # algorithms as ephemerismanager inherently sorts based on prns
+        idxs = np.where(navdata["gps_millis"] == timestep)[0]
+        sorted_idxs = idxs[np.argsort(navdata["sv_id", idxs], axis = 0)]
+
+        # compute ephem information using desired_sats, rxdatetime
+        desired_sats = [unique_gnss_id_str + str(int(i)).zfill(2) \
+                                           for i in navdata["sv_id", sorted_idxs]]
+        rxdatetime = datetime(1980, 1, 6, 0, 0, 0, tzinfo=timezone.utc) + \
+                     timedelta( seconds = (timestep) * 1e-3 )
+        ephem = get_time_cropped_rinex(rxdatetime, satellites = desired_sats)
+
+        # compute satellite position and velocity based on ephem and gps_time
+        # Transform satellite position to account for earth's rotation
+        get_sat_from_ephem = find_sv_states(timestep, ephem)
+        satpos_ephemeris = np.transpose([get_sat_from_ephem["x_sv_m"], \
+                                         get_sat_from_ephem["y_sv_m"], \
+                                         get_sat_from_ephem["z_sv_m"]])
+        satvel_ephemeris = np.transpose([get_sat_from_ephem["vx_sv_mps"], \
+                                         get_sat_from_ephem["vy_sv_mps"], \
+                                         get_sat_from_ephem["vz_sv_mps"]])
+        trans_time = navdata["raw_pr_m", sorted_idxs] / consts.C
+        del_x = (consts.OMEGA_E_DOT * satpos_ephemeris[:,1] * trans_time)
+        del_y = (-consts.OMEGA_E_DOT * satpos_ephemeris[:,0] * trans_time)
+        satpos_ephemeris[:,0] = satpos_ephemeris[:,0] + del_x
+        satpos_ephemeris[:,1] = satpos_ephemeris[:,1] + del_y
+
+        if verbose:
+            print('after ephemeris:', satpos_ephemeris, satvel_ephemeris)
+            satpos_android = np.transpose([ navdata["x_sv_m", sorted_idxs], \
+                                            navdata["y_sv_m", sorted_idxs], \
+                                            navdata["z_sv_m", sorted_idxs] ])
+            satvel_android = np.transpose([ navdata["vx_sv_mps", sorted_idxs], \
+                                               navdata["vy_sv_mps", sorted_idxs], \
+                                               navdata["vz_sv_mps", sorted_idxs] ])
+            print('nav-android Pos Error: ', \
+                      np.linalg.norm(satpos_ephemeris - satpos_android, axis=1) )
+            print('nav-android Vel Error: ', \
+                      np.linalg.norm(satvel_ephemeris - satvel_android, axis=1) )
+
+        # update *_sv_m of navdata with the estimated values from .n files
+        navdata["x_sv_m", sorted_idxs] = satpos_ephemeris[:,0]
+        navdata["y_sv_m", sorted_idxs] = satpos_ephemeris[:,1]
+        navdata["z_sv_m", sorted_idxs] = satpos_ephemeris[:,2]
+
+        # update v*_sv_mps of navdata with the estimated values from .n files
+        navdata["vx_sv_mps", sorted_idxs] = satvel_ephemeris[:,0]
+        navdata["vy_sv_mps", sorted_idxs] = satvel_ephemeris[:,1]
+        navdata["vz_sv_mps", sorted_idxs] = satvel_ephemeris[:,2]
+
+    return navdata
